Allow clients to check Hive error codes
Currently, when Hive errors are returned, they get collapsed down to an error string.

This change adds a wrapper struct - HiveError, that contains the usual error, as well as the ErrorCode from the TStatus for the response.

Clients can type assert the returned error and check the code, as is done in the added test.