fix: pass HiveConfiguration from DSN to connectConfiguration in OpenConnector (#259)
This fixes a bug where session configuration parameters (like Spark configs)
passed through the DSN were not being sent to the Hive/Kyuubi server.

The HiveConfiguration is correctly parsed by ParseDSN into dsn.HiveConfiguration,
and is used in hive.go line 316 (openSession.Configuration = configuration.HiveConfiguration),
but was not being passed from the DSN to the connectConfiguration in OpenConnector.

Fixes #258